updated all model eval files with experiments

+++ evaluate_owlv2.py
@@ -1,6 +1,8 @@
 from models.owlv2 import owl_v2
 from models.utils import *
 import json
+import os
+from time import time
 import torch
 from tqdm import tqdm
 
@@ -10,9 +12,10 @@ refcoco_annotation_path = "/data/yashowardhan/FineShot/data/refcoco data/refcoco
 vg_image_path = "/data/yashowardhan/FineShot/test/vg_images"
 vg_annotation_path = "/data/yashowardhan/FineShot/data/visual genome data/vg_subset.json"
 
-def inference(image_path, annotation_path, dataset="refcoco"):
+def inference(image_path, annotation_path, dataset="refcoco", confidence_threshold=0.5, test=False):
     image_paths = load_image_paths(image_path)
-    image_paths = image_paths[:5]
+    if test:
+        image_paths = image_paths[:100]
     annotations = load_annotations(annotation_path)
     
     results = {}
@@ -20,9 +23,9 @@ def inference(image_path, annotation_path, dataset="refcoco"):
     for image_path in tqdm(image_paths, desc="Inference using Owlv2"):
         try:
             if dataset == "refcoco":
-                detections = owl_v2(image_path, annotations[str(i)]["labels"], confidence_threshold=0.5)
+                detections = owl_v2(image_path, annotations[str(i)]["labels"], confidence_threshold=confidence_threshold)
             else:
-                detections = owl_v2(image_path, annotations[i]["labels"], confidence_threshold=0.5)
+                detections = owl_v2(image_path, annotations[i]["labels"], confidence_threshold=confidence_threshold)
             results[i] = detections
         except:
             print(f"Error in processing image {i}")
@@ -31,13 +34,31 @@ def inference(image_path, annotation_path, dataset="refcoco"):
 
 
 if __name__ == "__main__":
-    refcoco_results = inference(refcoco_image_path, refcoco_annotation_path, "refcoco")
-    with open("/data/yashowardhan/FineShot/test/results/owlv2_refcoco_results.json", "w") as f:
-        json.dump(refcoco_results, f, indent=4)
+    # Confidence thresholds
+    cts = [0.1, 0.5, 0.9]
+    for ct in cts:
+        print(f"Confidence Threshold: {ct}")
+        file_ct = str(ct).replace(".", "")
+        
+        print("Dataset: RefCOCO")
+        refcoco_result_file = f"/data/yashowardhan/FineShot/test/results/owlv2_{file_ct}_refcoco_results.json"
+        if not os.path.exists(refcoco_result_file):
+            start = time()
+            refcoco_results = inference(refcoco_image_path, refcoco_annotation_path, "refcoco", ct)
+            end = time()
+            print(f"Inference Time: {end-start} sec")
+            with open(refcoco_result_file, "w") as f:
+                json.dump(refcoco_results, f, indent=4)
+        
+        print("Dataset: VG")
+        vg_result_file = f"/data/yashowardhan/FineShot/test/results/owlv2_{file_ct}_vg_results.json"
+        if not os.path.exists(vg_result_file):
+            start = time()
+            vg_results = inference(vg_image_path, vg_annotation_path, "vg", ct)
+            end = time()
+            print(f"Inference Time: {end-start} sec")
+            with open(vg_result_file, "w") as f:
+                json.dump(vg_results, f, indent=4)
     
-    vg_results = inference(vg_image_path, vg_annotation_path, "vg")
-    with open("/data/yashowardhan/FineShot/test/results/owlv2_vg_results.json", "w") as f:
-        json.dump(vg_results, f, indent=4)
-        
     print("Results saved successfully!")
-        
+
